initial commit: things already done: -free cell game works -shuffle -drag/drop cards -card position visual interpolation -quick move cards by double clicking -cascading cards winning animation -hotkeys: r-restart (new game), w-auto win(to test winning animation) currently working on: -undo/redo system -simple bug fixes -graphical improvements
At this point, this game is almost finished, the last major thing to do is the undo/redo system, and then a bunch of little bugs/quality things I have written down. Currently, I have the shuffle turned off for testing, there should be a commented line of code in the game.odin file to turn it back on.